some bug fixed

+++ server/Router/User-router.js
@@ -3,40 +3,123 @@ const User = require('../Models/User_model');
 const Post = require('../Models/Post_model');
 const bcrypt = require('bcrypt');
 const fs = require("fs")
-const authgurd = require("../authgard/authgurd")
+const authgurd = require("../authgard/authgurd");
+const upload = require("../Multer/Multer");
+const path = require('path');
+const jwt = require('jsonwebtoken');
 
 // update account
+router.post('/:id', authgurd, upload.single('files'), async (req, res) => {
 
-router.put('/:id', authgurd, async (req, res) => {
+    const { username, email, password } = req.body;
 
-    if (req.body.userId === req.params.id) {
+    if (req.userId === req.params.id) {
 
-        if (req.body.profilepic) {
-            const user = await User.findById(req.body.userId)
+        const user = await User.findById(req.userId)
+
+        if (req.file) {
+
+            user.profilepic = req.file.filename;
+
             const oldPhoto = user.profilepic;
             const uploadDir = "upload/";
-            const oldPhotoWithPath = uploadDir + oldPhoto;
+            const oldPhotoWithPath = path.join(__dirname, uploadDir, oldPhoto);
 
-            if (fs.existsSync(oldPhotoWithPath)) {
-                fs.unlink(oldPhotoWithPath, (err) => {
-                    console.log(err);
-                });
+            if (oldPhoto) {
+
+                if (fs.existsSync(oldPhotoWithPath)) {
+                    fs.unlink(oldPhotoWithPath, (err) => {
+                        console.log(err);
+                    });
+                }
+
+            } else {
+
+                console.log("oldPhoto is not exist");
             }
-        }
 
-        if (req.body.password) {
-            req.body.password = await bcrypt.hash(req.body.password, 10);
-            try {
-                const updateduser = await User.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
-                res.status(200).json(updateduser);
-            } catch (err) { res.status(500).json(err); }
+            if (req.body.password) {
+
+                hassedPassword = await bcrypt.hash(req.body.password, 10);
+                user.password = hassedPassword;
+
+                try {
+
+                    username && (user.username = username);
+                    email && (user.email = email);
+                    await user.save();
+                    const updateduser = await User.findById(req.params.id);
+
+                    const { password, ...others } = updateduser._doc;
+                    const token = await jwt.sign({ username: user.username, userid: user._id }, process.env.TOKENSECRATE);
+                    res.status(200).json({ ...others, token });
+
+                } catch (err) {
+
+                    res.status(400).json(err);
+                }
+
+            } else {
+
+                try {
+
+                    username && (user.username = username);
+                    email && (user.email = email);
+                    await user.save();
+                    const updateduser = await User.findById(req.params.id)
+                    const { password, ...others } = updateduser._doc;
+                    const token = await jwt.sign({ username: user.username, userid: user._id }, process.env.TOKENSECRATE);
+                    res.status(200).json({ ...others, token });
+
+                } catch (err) {
+
+                    res.status(500).json(err);
+                }
+            }
 
         } else {
-            try {
-                const updateduser = await User.findByIdAndUpdate(req.params.id, { $set: req.body }, { new: true })
-                res.status(200).json(updateduser);
-            } catch (err) { res.status(500).json(err); }
+
+            if (req.body.password) {
+
+                hassedPassword = await bcrypt.hash(req.body.password, 10);
+                user.password = hassedPassword;
+
+                try {
+
+                    username && (user.username = username);
+                    email && (user.email = email);
+                    await user.save();
+                    const updateduser = await User.findById(req.params.id)
+                    const { password, ...others } = updateduser._doc;
+                    const token = await jwt.sign({ username: user.username, userid: user._id }, process.env.TOKENSECRATE);
+                    res.status(200).json({ ...others, token });
+
+                } catch (err) {
+
+                    res.status(400).json(err);
+                }
+
+            } else {
+
+                try {
+
+                    username && (user.username = username);
+                    email && (user.email = email);
+                    await user.save();
+
+                    const updateduser = await User.findById(req.params.id)
+                    const { password, ...others } = updateduser._doc;
+                    const token = await jwt.sign({ username: user.username, userid: user._id }, process.env.TOKENSECRATE);
+                    res.status(200).json({ ...others, token });
+
+                } catch (err) {
+
+                    res.status(500).json(err);
+                }
+            }
+
         }
+
     } else {
 
         res.status(400).json("you can update only your account!!");
@@ -47,8 +130,8 @@ router.put('/:id', authgurd, async (req, res) => {
 
 router.delete('/:id', authgurd, async (req, res) => {
 
+    if (req.userId === req.params.id) {
 
-    if (req.body.userId === req.params.id) {
         try {
             const user = await User.findById(req.params.id);
 
@@ -56,27 +139,39 @@ router.delete('/:id', authgurd, async (req, res) => {
                 if (user.profilepic) {
                     const oldPhoto = user.profilepic;
                     const uploadDir = "upload/";
-                    const oldPhotoWithPath = uploadDir + oldPhoto;
+                    const oldPhotoWithPath = path.join(__dirname, uploadDir, oldPhoto);
 
-                    if (fs.existsSync(oldPhotoWithPath)) {
-                        fs.unlink(oldPhotoWithPath, (err) => {
-                            if (err) { new Error("could not delete profile picture") }
-                        });
+                    if (oldPhoto) {
+
+                        if (fs.existsSync(oldPhotoWithPath)) {
+                            fs.unlink(oldPhotoWithPath, (err) => {
+                                console.log(err);
+                            });
+                        }
+
+                    } else {
+
+                        console.log("oldPhoto is not exist");
                     }
                 }
                 const allpost = await Post.find({ username: user.username })
                 allpost.map((post) => {
 
-                    if (post.photo) {
-                        const oldPhoto = post.photo;
-                        const uploadDir = "upload/";
-                        const oldPhotoWithPath = uploadDir + oldPhoto;
+                    const oldPhoto = user.profilepic;
+                    const uploadDir = "upload/";
+                    const oldPhotoWithPath = path.join(__dirname, uploadDir, oldPhoto);
+
+                    if (oldPhoto) {
 
                         if (fs.existsSync(oldPhotoWithPath)) {
                             fs.unlink(oldPhotoWithPath, (err) => {
-                                if (err) { new Error("Culd not delete photos") }
+                                console.log(err);
                             });
                         }
+
+                    } else {
+
+                        console.log("oldPhoto is not exist");
                     }
 
                 })
